+++ Source/AtCoderLibrary/STL/StlFunction.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
 
-namespace AtCoder
+namespace AtCoder.Stl
 {
     using static MethodImplOptions;
     public static class StlFunction